Harden Runtime Pack cross-platform contracts
