+++ .github/scripts/windows_test_assignments.json
@@ -520,6 +520,7 @@
       "tests/test_gateway/test_rpc_router_decisions.py",
       "tests/test_gateway/test_rpc_routing_hold.py",
       "tests/test_gateway/test_rpc_run_mode_preference.py",
+      "tests/test_gateway/test_rpc_sandbox_runtime.py",
       "tests/test_gateway/test_rpc_sandbox_setup.py",
       "tests/test_gateway/test_rpc_selflearning_status.py",
       "tests/test_gateway/test_rpc_session_services.py",
@@ -667,6 +668,9 @@
       "tests/test_readme_links.py",
       "tests/test_router_self_learning_dataset.py",
       "tests/test_router_self_learning_feedback.py",
+      "tests/test_runtime_packs/test_catalog.py",
+      "tests/test_runtime_packs/test_manager.py",
+      "tests/test_runtime_packs/test_shell_runtime_resolution.py",
       "tests/test_safety/test_secret_redaction.py",
       "tests/test_sandbox/test_channel_approval_gate.py",
       "tests/test_sandbox/test_elevation.py",
@@ -1244,7 +1248,9 @@
       "tests/test_scripts/test_build_wheelhouse_zip.py",
       "tests/test_scripts/test_exp_ledger.py",
       "tests/test_scripts/test_meta_skill_validation_matrix.py",
+      "tests/test_scripts/test_prestage_release_to_oss.py",
       "tests/test_scripts/test_release_channel_manifest.py",
+      "tests/test_scripts/test_verify_desktop_slim_size.py",
       "tests/test_scripts/test_verify_webui_artifact.py",
       "tests/test_skills/test_clarify_skip_summary.py",
       "tests/test_skills/test_creator_trigger_no_collision.py",
