+++ .github/scripts/windows_test_durations.json
@@ -487,6 +487,7 @@
     "tests/test_gateway/test_rpc_provider_credential_clear.py": 0.01,
     "tests/test_gateway/test_rpc_router_decisions.py": 14.966,
     "tests/test_gateway/test_rpc_routing_hold.py": 0.348,
+    "tests/test_gateway/test_rpc_sandbox_runtime.py": 0.01,
     "tests/test_gateway/test_rpc_sandbox_setup.py": 0.01,
     "tests/test_gateway/test_rpc_selflearning_status.py": 0.188,
     "tests/test_gateway/test_rpc_session_services.py": 0.01,
@@ -795,6 +796,9 @@
     "tests/test_router_self_learning_promotion.py": 0.576,
     "tests/test_router_self_learning_training.py": 0.863,
     "tests/test_router_tier_contract.py": 0.127,
+    "tests/test_runtime_packs/test_catalog.py": 0.01,
+    "tests/test_runtime_packs/test_manager.py": 0.01,
+    "tests/test_runtime_packs/test_shell_runtime_resolution.py": 0.01,
     "tests/test_safety/test_run_sandboxed_env.py": 0.01,
     "tests/test_safety/test_run_sandboxed_proc_cleanup.py": 0.01,
     "tests/test_safety/test_secret_redaction.py": 0.01,
@@ -920,10 +924,12 @@
     "tests/test_scripts/test_live_meta_soft_activation_e2e.py": 1.539,
     "tests/test_scripts/test_meta_skill_validation_matrix.py": 0.034,
     "tests/test_scripts/test_meta_trigger_accuracy_script.py": 0.012,
+    "tests/test_scripts/test_prestage_release_to_oss.py": 0.01,
     "tests/test_scripts/test_refresh_models_dev_snapshot.py": 0.011,
     "tests/test_scripts/test_release_channel_manifest.py": 0.01,
     "tests/test_scripts/test_replay_finalize_gate.py": 0.069,
     "tests/test_scripts/test_router_live_script_defaults.py": 0.038,
+    "tests/test_scripts/test_verify_desktop_slim_size.py": 0.01,
     "tests/test_scripts/test_verify_webui_artifact.py": 0.01,
     "tests/test_search/test_bocha_provider.py": 0.025,
     "tests/test_search/test_brave_config.py": 0.048,
